NN-4571 further refactoring
