+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/controllers/reported/HearingController.kt
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.HearingSummaryDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported.HearingService
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported.ReportedAdjudicationService
 import java.time.LocalDate
 import java.time.LocalDateTime
 
@@ -41,7 +40,6 @@ data class HearingRequest(
 @RestController
 class HearingController(
   val hearingService: HearingService,
-  val reportedAdjudicationService: ReportedAdjudicationService,
 ) : ReportedAdjudicationBaseController() {
 
   @PostMapping(value = ["/{adjudicationNumber}/hearing"])
@@ -98,7 +96,7 @@ class HearingController(
     @PathVariable(name = "agencyId") agencyId: String,
     @RequestParam(name = "hearingDate") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) hearingDate: LocalDate,
   ): HearingSummaryResponse {
-    val hearings = reportedAdjudicationService.getAllHearingsByAgencyIdAndDate(agencyId, hearingDate)
+    val hearings = hearingService.getAllHearingsByAgencyIdAndDate(agencyId, hearingDate)
 
     return HearingSummaryResponse(
       hearings
