+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/controllers/reported/ReportedAdjudicationControllerTest.kt
@@ -1,18 +1,12 @@
 package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.reported
 
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentMatchers.anyLong
-import org.mockito.kotlin.any
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest
 import org.springframework.boot.test.mock.mockito.MockBean
-import org.springframework.data.domain.PageImpl
-import org.springframework.data.domain.PageRequest
-import org.springframework.data.domain.Pageable
-import org.springframework.data.domain.Sort
 import org.springframework.security.test.context.support.WithMockUser
 import org.springframework.test.web.servlet.ResultActions
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
@@ -22,9 +16,6 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.TestControllerBase
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedAdjudicationStatus
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported.ReportedAdjudicationService
-import java.time.LocalDate
-import java.time.LocalDateTime
-import java.util.Optional
 import javax.persistence.EntityNotFoundException
 
 @WebMvcTest(value = [ReportedAdjudicationController::class])
@@ -67,125 +58,6 @@ class ReportedAdjudicationControllerTest : TestControllerBase() {
       return mockMvc
         .perform(
           get("/reported-adjudications/$adjudicationNumber")
-            .header("Content-Type", "application/json")
-        )
-    }
-  }
-
-  @Nested
-  inner class MyReportedAdjudications {
-    @BeforeEach
-    fun beforeEach() {
-      whenever(
-        reportedAdjudicationService.getMyReportedAdjudications(
-          any(), any(), any(), any(), any()
-        )
-      ).thenReturn(
-        PageImpl(
-          listOf(REPORTED_ADJUDICATION_DTO),
-          Pageable.ofSize(20).withPage(0),
-          1
-        ),
-      )
-    }
-
-    @Test
-    fun `responds with a unauthorised status code`() {
-      getMyAdjudications().andExpect(status().isUnauthorized)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER")
-    fun `makes a call to return my reported adjudications`() {
-      getMyAdjudications().andExpect(status().isOk)
-      verify(reportedAdjudicationService).getMyReportedAdjudications(
-        "MDI",
-        LocalDate.now().minusDays(3), LocalDate.now(), Optional.empty(),
-        PageRequest.ofSize(20).withPage(0).withSort(
-          Sort.by(
-            Sort.Direction.DESC,
-            "incidentDate"
-          )
-        )
-      )
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["ROLE_ADJUDICATIONS_REVIEWER"])
-    fun `makes a call to return all reported adjudications`() {
-      getAllAdjudications().andExpect(status().isOk)
-      verify(reportedAdjudicationService).getAllReportedAdjudications(
-        "MDI",
-        LocalDate.now().minusDays(3), LocalDate.now(), Optional.empty(),
-        PageRequest.ofSize(20).withPage(0).withSort(
-          Sort.by(
-            Sort.Direction.DESC,
-            "incidentDate"
-          )
-        )
-      )
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER")
-    fun `returns my reported adjudications`() {
-      getMyAdjudications()
-        .andExpect(status().isOk)
-        .andExpect(jsonPath("$.totalPages").value(1))
-        .andExpect(jsonPath("$.size").value(20))
-        .andExpect(jsonPath("$.number").value(0))
-        .andExpect(jsonPath("$.content[0].adjudicationNumber").value(1))
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER")
-    fun `returns my reported adjudications with date and status filter`() {
-      getMyAdjudicationsWithFilter(LocalDate.now().plusDays(5), ReportedAdjudicationStatus.AWAITING_REVIEW)
-        .andExpect(status().isOk)
-        .andExpect(jsonPath("$.totalPages").value(1))
-        .andExpect(jsonPath("$.size").value(20))
-        .andExpect(jsonPath("$.number").value(0))
-        .andExpect(jsonPath("$.content[0].adjudicationNumber").value(1))
-
-      verify(reportedAdjudicationService).getMyReportedAdjudications(
-        "MDI",
-        LocalDate.now().plusDays(5),
-        LocalDate.now().plusDays(5),
-        Optional.of(ReportedAdjudicationStatus.AWAITING_REVIEW),
-        PageRequest.ofSize(20).withPage(0).withSort(
-          Sort.by(
-            Sort.Direction.DESC,
-            "incidentDate"
-          )
-        )
-      )
-    }
-
-    @Test
-    fun `paged responds with a unauthorised status code`() {
-      getMyAdjudications().andExpect(status().isUnauthorized)
-    }
-
-    private fun getMyAdjudications(): ResultActions {
-      return mockMvc
-        .perform(
-          get("/reported-adjudications/my/agency/MDI?page=0&size=20&sort=incidentDate,DESC")
-            .header("Content-Type", "application/json")
-        )
-    }
-
-    private fun getAllAdjudications(): ResultActions {
-      return mockMvc
-        .perform(
-          get("/reported-adjudications/agency/MDI?page=0&size=20&sort=incidentDate,DESC")
-            .header("Content-Type", "application/json")
-        )
-    }
-
-    private fun getMyAdjudicationsWithFilter(date: LocalDate, reportedAdjudicationStatus: ReportedAdjudicationStatus): ResultActions {
-      return mockMvc
-        .perform(
-          get("/reported-adjudications/my/agency/MDI?startDate=$date&endDate=$date&status=$reportedAdjudicationStatus&page=0&size=20&sort=incidentDate,DESC")
             .header("Content-Type", "application/json")
         )
     }
@@ -247,11 +119,5 @@ class ReportedAdjudicationControllerTest : TestControllerBase() {
         status().isUnauthorized
       )
     }
-  }
-
-  companion object {
-    private val DATE_TIME_OF_INCIDENT = LocalDateTime.of(2010, 10, 12, 10, 0, 0)
-    private val DATE_TIME_DRAFT_ADJUDICATION_HANDOVER_DEADLINE = LocalDateTime.of(2010, 10, 14, 10, 0)
-    private const val INCIDENT_STATEMENT = "A statement"
   }
 }
