+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/ReportsServiceTest.kt
@@ -1,0 +1,143 @@
+package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported
+
+import org.assertj.core.api.Java6Assertions.assertThat
+import org.assertj.core.groups.Tuple
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+import org.mockito.kotlin.any
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
+import org.springframework.data.domain.PageImpl
+import org.springframework.data.domain.Pageable
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedAdjudicationStatus
+import java.time.LocalDate
+import java.time.LocalTime
+import java.util.Optional
+
+class ReportsServiceTest : ReportedAdjudicationTestBase() {
+
+  private val reportsService = ReportsService(
+    reportedAdjudicationRepository, authenticationFacade, offenceCodeLookupService
+  )
+
+  @Nested
+  inner class AllReportedAdjudications {
+    @BeforeEach
+    fun beforeEach() {
+      val reportedAdjudication1 =
+        entityBuilder.reportedAdjudication(reportNumber = 1L, dateTime = DATE_TIME_OF_INCIDENT)
+      reportedAdjudication1.createdByUserId = "A_SMITH"
+      reportedAdjudication1.createDateTime = REPORTED_DATE_TIME
+      val reportedAdjudication2 =
+        entityBuilder.reportedAdjudication(reportNumber = 2L, dateTime = DATE_TIME_OF_INCIDENT)
+      reportedAdjudication2.createdByUserId = "P_SMITH"
+      reportedAdjudication2.createDateTime = REPORTED_DATE_TIME.plusDays(2)
+      whenever(
+        reportedAdjudicationRepository.findByAgencyIdAndDateTimeOfDiscoveryBetweenAndStatusIn(
+          any(),
+          any(),
+          any(),
+          any(),
+          any()
+        )
+      ).thenReturn(
+        PageImpl(
+          listOf(reportedAdjudication1, reportedAdjudication2)
+        )
+      )
+    }
+
+    @Test
+    fun `makes a call to the reported adjudication repository to get the page of adjudications`() {
+      reportsService.getAllReportedAdjudications(
+        "MDI",
+        LocalDate.now(),
+        LocalDate.now(),
+        Optional.empty(),
+        Pageable.ofSize(20).withPage(0)
+      )
+
+      verify(reportedAdjudicationRepository).findByAgencyIdAndDateTimeOfDiscoveryBetweenAndStatusIn(
+        "MDI",
+        LocalDate.now().atStartOfDay(),
+        LocalDate.now().atTime(LocalTime.MAX),
+        ReportedAdjudicationStatus.values().toList(),
+        Pageable.ofSize(20).withPage(0)
+      )
+    }
+
+    @Test
+    fun `returns all reported adjudications`() {
+      val myReportedAdjudications = reportsService.getAllReportedAdjudications(
+        "MDI",
+        LocalDate.now(),
+        LocalDate.now(),
+        Optional.empty(),
+        Pageable.ofSize(20).withPage(0)
+      )
+
+      assertThat(myReportedAdjudications.content)
+        .extracting("adjudicationNumber", "prisonerNumber", "bookingId", "createdByUserId", "createdDateTime")
+        .contains(
+          Tuple.tuple(1L, "A12345", 234L, "A_SMITH", REPORTED_DATE_TIME),
+          Tuple.tuple(2L, "A12345", 234L, "P_SMITH", REPORTED_DATE_TIME.plusDays(2))
+        )
+    }
+  }
+
+  @Nested
+  inner class MyReportedAdjudications {
+    @BeforeEach
+    fun beforeEach() {
+      val reportedAdjudication1 =
+        entityBuilder.reportedAdjudication(reportNumber = 1L, dateTime = DATE_TIME_OF_INCIDENT)
+      reportedAdjudication1.createdByUserId = "A_SMITH"
+      reportedAdjudication1.createDateTime = REPORTED_DATE_TIME
+      val reportedAdjudication2 =
+        entityBuilder.reportedAdjudication(reportNumber = 2L, dateTime = DATE_TIME_OF_INCIDENT)
+      reportedAdjudication2.createdByUserId = "P_SMITH"
+      reportedAdjudication2.createDateTime = REPORTED_DATE_TIME.plusDays(2)
+      whenever(
+        reportedAdjudicationRepository.findByCreatedByUserIdAndAgencyIdAndDateTimeOfDiscoveryBetweenAndStatusIn(
+          any(),
+          any(),
+          any(),
+          any(),
+          any(),
+          any()
+        )
+      ).thenReturn(
+        PageImpl(
+          listOf(reportedAdjudication1, reportedAdjudication2)
+        )
+      )
+    }
+
+    @Test
+    fun `returns my reported adjudications`() {
+      val myReportedAdjudications = reportsService.getMyReportedAdjudications(
+        "MDI",
+        LocalDate.now(),
+        LocalDate.now(),
+        Optional.empty(),
+        Pageable.ofSize(20).withPage(0)
+      )
+
+      assertThat(myReportedAdjudications.content)
+        .extracting("adjudicationNumber", "prisonerNumber", "bookingId", "createdByUserId", "createdDateTime")
+        .contains(
+          Tuple.tuple(1L, "A12345", 234L, "A_SMITH", REPORTED_DATE_TIME),
+          Tuple.tuple(2L, "A12345", 234L, "P_SMITH", REPORTED_DATE_TIME.plusDays(2))
+        )
+    }
+  }
+
+  override fun `throws an entity not found if the reported adjudication for the supplied id does not exists`() {
+    // not required.
+  }
+
+  companion object {
+    private val REPORTED_DATE_TIME = DATE_TIME_OF_INCIDENT.plusDays(1)
+  }
+}
